Fix url and secret key; move to a single file
